refactor: mission_service에서 shared preferences 의존성 제거
